✨ feat: 메인페이지 기능 구현(메인페이지 경로 변경, 메인페이지 내 기능 구현 코드 포스트카드로 옮기기)
