+++ src/pages/MainPage/components/MainSmallBanner.vue
@@ -1,6 +1,4 @@
 <script setup>
-import { defineProps } from 'vue';
-
 defineProps({
   to: String,
   bgColor: String,
